renamed support folder to styles
