+++ client/src/App.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import MainContainer from './connected-components/MainContainer';
 import logo from './support/small-header.svg';
-import './support/App.css';
+import './styles/App.css';
 
 class App extends Component {
   render() {
