+++ client/src/styles/DetailPane.css
